Added cast due to compiler warning.
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@726 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSTravelcostDetector.h
@@ -21,6 +21,9 @@
  ***************************************************************************/
 
 // $Log$
+// Revision 1.3  2003/06/06 11:00:35  roessel
+// Added cast due to compiler warning.
+//
 // Revision 1.2  2003/06/05 12:51:26  roessel
 // Modified #includes and changed getString() to toString().
 //
@@ -108,7 +111,7 @@ public:
                 ( this, &MSTravelcostDetector::write2file, intervalLength );
             MSEventControl::getEndOfTimestepEvents()->addEvent(
                 writeData,
-                intervalLength / MSNet::deltaT() - 1,
+                static_cast<int>( intervalLength / MSNet::deltaT() ) - 1,
                 MSEventControl::ADAPT_AFTER_EXECUTION );
         }
